MYNEWT-448: Add hal watchdog for stm32f3
NOTE: there is one issue unresolved with this commit. There is no
way to exit the loop where we are waiting for the register
changes to take effect. Not sure the best way to do this at this
point so for now we just skip it. This has been noted in the ticket
and should be addressed soon.